Desktop/Rest_Project/requirements.txt
django
djangorestframework
itsdangerous
jwt
